Move ensemble into core namespace
