+++ relentless/__init__.py
@@ -4,8 +4,6 @@ assert sys.version_info >= (3,4), "relentless requires Python 3.4"
 from . import core
 from .core import *
 from . import engine
-from . import ensemble
-from .ensemble import *
 from . import environment
 from . import optimize
 from . import potential
